requirements.txt
ansi2html==1.9.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
bootstrap4==0.1.0
certifi==2023.11.17
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
dash==2.9.3
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==4.2.9
django-bootstrap4==23.4
django-plotly-dash==2.2.2
dpd-components==0.1.0
Flask==3.0.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
nest-asyncio==1.5.9
numpy==1.24.4
packaging==23.2
pandas==2.0.3
plotly==5.18.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
retrying==1.3.4
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tenacity==8.2.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
ansi2html==1.9.1
asgiref==3.7.2
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
bootstrap4==0.1.0
certifi==2023.11.17
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
dash==2.9.3
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
Django==4.2.9
django-bootstrap-icons==0.8.7
django-bootstrap4==23.4
django-extensions==3.2.3
django-filter==24.2
django-plotly-dash==2.2.2
django-restframework==0.0.1
djangorestframework==3.15.1
dpd-components==0.1.0
Flask==3.0.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
Markdown==3.6
MarkupSafe==2.1.3
nest-asyncio==1.5.9
numpy==1.24.4
packaging==23.2
pandas==2.0.3
plotly==5.18.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
retrying==1.3.4
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tenacity==8.2.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
zipp==3.17.0
